auth-service/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==2.0.15
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pydantic==1.10.7
python-jose==3.3.0
passlib==1.7.4
tenacity==8.2.2
httpx==0.24.1
alembic==1.11.1
pytest==7.3.1
pytest-cov==4.1.0

#prometheus-client==0.17.0
#prometheus-fastapi-instrumentator==7.1.0

prometheus-fastapi-instrumentator==5.7.0
prometheus-client==0.8.0  # Explicitly set compatible version
python-json-logger==2.0.7
loguru==0.7.0
requests==2.31.0
python-multipart==0.0.20